feat(spx-gui): add caching wrapper for asset2model conversion
Introduced a caching wrapper to enhance performance
by caching the converted model(a sprite, backdrop, or sound) of an asset.

It could avoid redundant network requests (to fetch real asset data from kodo)
when reloading asset items in the virtual list.

Update asset prop of Asset Item to be a specific asset type