refactor(spx-gui): remove `select and add asset` functionality from AssetLibrary
Replaced the `Add to project` functionality with an `Add to project` button in the AssetItem component. (Done in the previous commit)
This change eliminates the need to select multiple assets and confirm the selection in the AssetLibraryModal.

The click event on the AssetItem component will trigger the `open asset details` functionality. (Work in progress)