fix(spx-gui): prevent race condition in asset search
Currently, the asset search is not working properly when the user changes the search params quickly.
The outdated search results are displayed to the user because
AssetList component just listens to the latest search results
and does not know whether the search results are outdated or not.
And thus the outdated search results will append to asset list.

This commit fixes the issue by adding a request id to useQuery hook.
Only the search results with the latest request id will be updated to `data`.
(This feature is disabled by default for compatibility with the current codebase.)