feat: Integrate ispx wasm
This commit adds a new file `ispxLoader.ts` that initializes the ispx wasm module
by creating an invisible iframe and loading the wasm module in it.
The wasm module is loaded from the `main.wasm` file located in the `@/assets/ispx` directory.
The iframe is hidden to prevent the wasm from adding a canvas to the document body.

The commit also added function `goParseSkeletonAnimData` and related types to call the
`goParseSkeletonAnimData` in the iframe.