feat: implements skeleton animation renderer
This commit implements the skeleton animation renderer for sprite.
The renderer is based on WebGL and uses the shader to render the skeleton animation.

The renderer accepts the skeleton animation data that is parsed from the SPX WASM and
contains the uv, vertices, and indices of each frame of the skeleton animation.

The renderer also provides the ability to control the playback and fps of the animation.

The `SkeletonAnimationRenderer.vue` is the Vue component that renders the skeleton animation.
The `shader.vert` and `shader.frag` are the vertex and fragment shaders for the skeleton animation renderer.