refactor: flatten vector array before transfer to js
The vert anim and skeleton anim take a large number of vertices and uvs
to render. Currently, the parsed vertices and uvs from the ispx wasm
module are `[]math32.Vector3` and `[]math32.Vector2` respectively. This
large object array takes a long time to transfer to the js side,
whether we transfer via json (about 4s) or `js.ValueOf` (about 2s).

To improve performance, we can flatten the vertices and uvs into a
single typed array before transferring to the js side. We can directly
convert the whole array to `[]byte` and
do not need to convert each vector to js object

This commit convert the `[]math32.Vector3` or `[]math32.Vector2` to
`[]float32` and then `[]byte` before transferring to the js side.
The js runtime could then use `Uint8Array` and `numComponents` to
build buffer infos. Now the transfer time is reduced to about 300~400ms.