refactor: move content generation of AI asset into specific editor
The further AI generation used to be done in the AIPreviewModal component,
and the well-generated content was then passed to the specific editor components,
which is not the best place for it as we need to do some edits before the generation for sprite assets.

This commit moves the generation logic into the specific editor components.
In the following commits, we will make the sprite generation manual and add the inpaint feature.