feat: Add asset to Milvus database when adding asset to asset library
This change adds functionality to add the asset to Milvus.
The asset's vector embedding is obtained using the prompt and added to the Milvus database
along with the asset ID and asset name.

The AddAssetParams struct now includes a new optional field called "prompt".
This field represents the keyword used when generating the asset.
If the prompt is not provided, the display name of the asset will be used as the prompt.

These changes were made to make sure that the suggestions api has data to retrieve from the Milvus database
to provide search suggestions.